[DEP] Remove deprecated stats (#2963)
* Remove deprecated stats

* fixes and docs

* import fixes

* doctests

* alpha docs
