+++ aeon/benchmarking/metrics/anomaly_detection/__init__.py
@@ -14,9 +14,6 @@ __all__ = [
     "range_pr_auc_score",
     "range_pr_vus_score",
     "range_roc_vus_score",
-    "ts_precision",
-    "ts_recall",
-    "ts_fscore",
 ]
 
 from aeon.benchmarking.metrics.anomaly_detection._continuous import (
@@ -30,11 +27,6 @@ from aeon.benchmarking.metrics.anomaly_detection._range_metrics import (
     range_f_score,
     range_precision,
     range_recall,
-)
-from aeon.benchmarking.metrics.anomaly_detection._range_ts_metrics import (
-    ts_fscore,
-    ts_precision,
-    ts_recall,
 )
 from aeon.benchmarking.metrics.anomaly_detection._vus_metrics import (
     range_pr_auc_score,
