+++ docs/api_reference/benchmarking.md
@@ -1,17 +1,14 @@
-.. _benchmarking_ref:
+# Benchmarking
 
-Benchmarking
-============
-
-The ``aeon.benchmarking`` module contains tools for comparing and evaluating time
+The `aeon.benchmarking` module contains tools for comparing and evaluating time
 series models, loading stored results, and calculating performance metrics for a
 variety of tasks.
 
-Results loading
----------------
+## Results loading
 
-Results loaders and loading utilities for ``aeon`` (and other) estimators.
+Results loaders and loading utilities for `aeon` (and other) estimators.
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.results_loaders
 
 .. autosummary::
@@ -22,12 +19,13 @@ Results loaders and loading utilities for ``aeon`` (and other) estimators.
     get_available_estimators
     get_estimator_results
     get_estimator_results_as_array
+```
 
-Published results
-^^^^^^^^^^^^^^^^^
+## Published results
 
 Results loaders for specific publications.
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.published_results
 
 .. autosummary::
@@ -37,12 +35,13 @@ Results loaders for specific publications.
     load_classification_bake_off_2017_results
     load_classification_bake_off_2021_results
     load_classification_bake_off_2023_results
+```
 
-Resampling
-----------
+## Resampling
 
 Functions for resampling time series data.
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.resampling
 
 .. autosummary::
@@ -53,15 +52,15 @@ Functions for resampling time series data.
     resample_data_indices
     stratified_resample_data
     stratified_resample_data_indices
+```
 
-Performance metrics
--------------------
+## Performance metrics
 
-Performance metrics used for evaluating ``aeon`` estimators.
+Performance metrics used for evaluating `aeon` estimators.
 
-Anomaly Detection
-^^^^^^^^^^^^^^^^^
+### Anomaly Detection
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.metrics.anomaly_detection
 
 .. autosummary::
@@ -81,10 +80,11 @@ Anomaly Detection
     range_pr_auc_score
     range_pr_vus_score
     range_roc_vus_score
+```
 
-Anomaly detection thresholding
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+### Anomaly detection thresholding
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.metrics.anomaly_detection.thresholding
 
 .. autosummary::
@@ -95,10 +95,11 @@ Anomaly detection thresholding
     sigma_threshold
     top_k_points_threshold
     top_k_ranges_threshold
+```
 
-Clustering
-^^^^^^^^^^
+### Clustering
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.metrics.clustering
 
 .. autosummary::
@@ -106,10 +107,11 @@ Clustering
     :template: function.rst
 
     clustering_accuracy_score
+```
 
-Segmentation
-^^^^^^^^^^^^
+### Segmentation
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.metrics.segmentation
 
 .. autosummary::
@@ -119,10 +121,11 @@ Segmentation
     count_error
     hausdorff_error
     prediction_ratio
+```
 
-Stats
------
+## Stats
 
+```{eval-rst}
 .. currentmodule:: aeon.benchmarking.stats
 
 .. autosummary::
@@ -132,3 +135,4 @@ Stats
     check_friedman
     nemenyi_test
     wilcoxon_test
+```
